feat(i2c): rearrange module, split into pads, structure and register
